token 공통 함수로 빼기
