+++ views/common.py
@@ -10,6 +10,20 @@ app = Flask(__name__)
 app.secret_key = 'sparta'
 
 
+# import
+def check_decode():
+    token_receive = request.cookies.get('mytoken')
+    if token_receive is not None:
+        try:
+            payload = jwt.decode(token_receive, app.secret_key, algorithms=['HS256'])
+
+            return payload['user_id']
+        except:
+            return None
+    else:
+        return None
+
+
 @bp.route('/token_check', methods=['GET'])
 def token_check():
     token_receive = request.cookies.get('mytoken')
@@ -19,7 +33,6 @@ def token_check():
             return jsonify({'user_id': payload['user_id']})
         else:
             return jsonify({'user_id': 'None'})
-
     except jwt.ExpiredSignatureError:
         return redirect(url_for("sign_in", msg="로그인 시간이 만료되었습니다."))
     except jwt.exceptions.DecodeError:
